requirements.txt
certifi==2020.6.20
cycler==0.10.0
jieba==0.42.1
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.3
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
wincertstore==0.2
wordcloud==1.8.1
